GiveTimes now functional
